Add event logging and API response
The Lambda function now logs actions performed in JSON format and
returns it as part of an appropriately formed response from AWS Lambda
to the API Gateway.

In the future, consider the option to upload the event log to S3; while
the log may be useful for testing, it is useless as a reply to a
SimpleMDM webhook.